Move Coturn IceUser to coturn module
